backend/otif-genai/requirements.txt
pandas
openpyxl
openai
python-dotenv
tenacity